Made small changes to styling of the content, also changed the favicon
