+++ src/components/About.css
@@ -1,5 +1,5 @@
 .about-page {
-  padding: 2rem 1rem 5rem;
+  padding: 2rem 2rem 5rem;
 }
 
 .inline-block-text {
@@ -22,7 +22,10 @@
   padding-bottom: 2rem;
 }
 
-.about-description,
+.about-description {
+  text-align: left;
+}
+
 .about-contact-info {
   text-align: center;
 }
@@ -71,6 +74,10 @@
   font-size: 1rem;
 }
 
+.about-description {
+  padding-bottom: 4rem;
+}
+
 @media only screen and (min-width: 40rem) {
   .about-page-intro {
     font-size: 4rem;
@@ -91,7 +98,6 @@
   }
 
   .about-description {
-    padding: 0 5rem;
     grid-column: 1 / 9;
     grid-row: 3 / 4;
   }
@@ -125,9 +131,9 @@
     padding-bottom: 4rem;
   }
 
-  .about-description {
+  /* .about-description {
     padding: 0 15rem;
-  }
+  } */
 
   .about-page {
     grid-template-columns: repeat(12, 1fr);
@@ -139,7 +145,7 @@
   }
 
   .about-description {
-    grid-column: 1 / 13;
+    grid-column: 2 / 11;
     grid-row: 3 / 4;
   }
 
@@ -154,10 +160,9 @@
   }
 
   .about-profile-image {
-    max-width: 70rem;
-    max-height: 45rem;
+    aspect-ratio: 16/8;
     margin: 0 auto;
-    margin-bottom: 3rem;
+    margin-bottom: 10rem;
     object-fit: cover;
   }
 }
