+++ src/components/Home.css
@@ -24,7 +24,7 @@
 }
 
 .landing-page {
-  height: 48rem;
+  height: 100vh;
   position: relative;
 }
