+++ src/components/SelectedWorks.css
@@ -1,7 +1,7 @@
 .selected-works-wrapper {
   /* max-width: 1200px; */
   width: 100%;
-  height: 100%;
+  height: 100vh;
   margin: 0 auto;
   padding: 2rem 2rem 7rem;
 }
