--- src/components/utilities/MusicPlayer.css
@@ -1,183 +1,0 @@
-.audio-button {
-  border: none;
-  color: #542316;
-  background-color: transparent;
-}
-
-.audio-player {
-  position: fixed;
-  width: 98%;
-  border: solid 0.05rem;
-  border-color: #542316;
-  background-color: #f7f3f0;
-}
-
-.progress-bar {
-  --bar-bg: #542316;
-
-  appearance: none;
-  background: var(--bar-bg);
-  border-radius: 1rem;
-  position: relative;
-  width: 100%;
-  height: 0.3rem;
-  outline: none;
-  cursor: pointer;
-}
-
-/* progress bar - safari */
-.progress-bar::-webkit-slider-runnable-track {
-  background: var(--bar-bg);
-  border-radius: 1rem;
-  position: relative;
-
-  width: 100%;
-  height: 0.3rem;
-  cursor: pointer;
-}
-
-/* progress bar - chrome and safari */
-.progress-bar::before {
-  content: "";
-  width: 100%;
-  height: 0.3rem;
-  background-color: var(--bar-bg);
-  border-radius: 1rem;
-  outline: none;
-  cursor: pointer;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 2;
-}
-
-/* knobby - chrome and safari */
-.progress-bar::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  height: 0.7rem;
-  width: 0.7rem;
-  border-radius: 50%;
-  border: none;
-  background-color: var(--bar-bg);
-  position: relative;
-  z-index: 3;
-  margin: -3px 0 0 0;
-}
-
-.progress-bar:active::-webkit-slider-thumb {
-  transform: scale(1.3);
-}
-
-.audio-player {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: auto;
-  column-gap: 1rem;
-  padding: 0.5rem 1rem;
-  margin: 2rem 1%;
-}
-
-.audio-progress-bar {
-  display: flex;
-  align-items: center;
-  padding: 1rem 0;
-}
-
-.audio-button {
-  display: flex;
-  align-items: center;
-}
-
-.audio-time-before-after {
-  padding-right: 1rem;
-}
-
-.song-name {
-  grid-column: 1 / 3;
-  grid-row: 1 / 2;
-}
-
-.audio-progress-bar {
-  grid-column: 1 / 5;
-  grid-row: 2 / 3;
-}
-
-.audio-controls {
-  grid-column: 3 / 5;
-  grid-row: 1 / 2;
-}
-
-.song-name,
-.audio-time-before-after {
-  font-size: 0.7rem;
-}
-
-.audio-controls,
-.audio-control-buttons,
-.song-name {
-  display: flex;
-  align-items: center;
-}
-
-.audio-controls {
-  justify-content: flex-end;
-}
-
-@media only screen and (min-width: 50rem) {
-  .audio-controls,
-  .audio-control-buttons,
-  .song-name {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .audio-player {
-    grid-template-columns: repeat(8, 1fr);
-    margin: 2rem 1%;
-  }
-
-  .audio-progress-bar {
-    display: flex;
-    align-items: center;
-    padding: 1rem 0;
-  }
-
-  .song-name {
-    grid-column: 1 / 2;
-    grid-row: 1 / 2;
-  }
-
-  .audio-progress-bar {
-    grid-column: 2 / 7;
-    grid-row: 1 / 2;
-  }
-
-  .audio-controls {
-    grid-column: 7 / 9;
-    grid-row: 1 / 2;
-  }
-}
-
-@media only screen and (min-width: 70rem) {
-  .audio-player {
-    grid-template-columns: repeat(12, 1fr);
-    padding: 0.5rem 1rem;
-    margin: 2rem 1%;
-  }
-
-  .song-name {
-    grid-column: 1 / 2;
-    grid-row: 1 / 2;
-  }
-
-  .audio-progress-bar {
-    grid-column: 2 / 11;
-    grid-row: 1 / 2;
-  }
-
-  .audio-controls {
-    grid-column: 11 / 13;
-    grid-row: 1 / 2;
-  }
-}
