+++ src/index.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: TRJND;
-  src: url(./assets/fonts/TRJND-font/TRJNDaVinci-Regular-Trial.ttf);
+  src: url(./assets/fonts/TRJND-font/TRJNDaVinci-Light-Trial.ttf);
 }
 
 * {
